Seperate files into multiple packages
