+++ src/main/java/aegis/exception/AegisException.java
@@ -1,0 +1,17 @@
+package aegis.exceptions;
+
+/**
+ * A base exception class for the Aegis chatbot.
+ * All custom exceptions in the Aegis chatbot extend this class.
+ */
+public class AegisException extends Exception {
+    /**
+     * Constructs an aegis.exceptions.AegisException with the specified error message.
+     *
+     * @param message The error message.
+     */
+    public AegisException(String message) {
+        super(message);
+    }
+}
+
