+++ src/main/java/aegis/exception/FileSavingException.java
@@ -1,0 +1,10 @@
+package aegis.exceptions;
+
+public class FileSavingException extends AegisException {
+    public FileSavingException(String message) {super(message);}
+
+    @Override
+    public String toString() {
+        return "Error Saving File: " + super.getMessage();
+    }
+}
