+++ src/main/java/aegis/task/Deadline.java
@@ -1,3 +1,5 @@
+import aegis.exceptions.TaskInputException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
