+++ src/main/java/aegis/ui/UIManager.java
